third_party/blink/web_tests/flag-specific/highdpi/virtual/composite-after-paint/scrollingcoordinator/donot-compute-non-fast-scrollable-region-for-hidden-frames-expected.txt

This is a testharness.js-based test.
FAIL This test ensures that non-fast scrollable area is re-computed on FrameView visibility changes. assert_equals: expected "[51, 102, 200, 200]" but got "[77, 1053, 333, 333]"
Harness: the test ran to completion.
